tools/rkn: download registry when it's updated
A registry is downloaded on every rknr_get.pl run. This is not
suitable in cases where urgent entries are added into the registry
and we must apply it in a hour interval.

To solve this problem, we can:

1. tell cron to call rknr_get.sh every 30 minutes
2. or, call rknr_get.sh every 5-10 minutes, but download the
   registry only if last update time of urgent entries is changed.

So, 2 variant is choosen.

rknr_get.pl is download the registry if:

1. last update time of urgent entries is newer than last download time
2. or, time elapsed from a last download if bigger than MAX_DOWNLOAD_INTERVAL