tools/rkn: change time delay logic a little
Wait before a 1 try a 1.5 minute to give a service a time to
prepare a requested registry copy. Then, enlarge a time delay
between tries by 5 secods after each try(initial value of a time delay
is 20 seconds; so, the time delay after a first try is 25 seconds).
After we reach a 300 seconds of a time delay, we reset it to 25 seconds.